PuzzlePrefabs(notVR ready)
to use: import the folder you want into unity and place the PrefabVariant in the project, currently, they work from keyboard inputs (check script to see which).
Conversion to VR can be done using onCollisionEnter of the hand in the trigger colliders of each button. The buttons are placed in the model but aren't mapped/labeled to actions. For the script it is only required to make a bool true for half second and make it false after (easy with coroutines) so that the Update function will see the input and proceed with the functionality of the puzzle.